[refactor] Home_컴포넌트 폴더로 분리 및 아이콘 정리
