+++ src/apis/post.ts
@@ -1,65 +1,65 @@
 import { axiosInstance } from "./axios";
 
 export const getPosts = async (channelId: string) => {
-  try {
-    const { data } = await axiosInstance.get(`/posts/channel/${channelId}`);
-    return data;
-  } catch (error) {
-    if (error instanceof Error) {
-      console.log(error.message);
-    } else {
-      console.log("Unknwon Error", error);
-    }
-  }
+	try {
+		const { data } = await axiosInstance.get(`/posts/channel/${channelId}`);
+		return data;
+	} catch (error) {
+		if (error instanceof Error) {
+			console.log(error.message);
+		} else {
+			console.log("Unknwon Error", error);
+		}
+	}
 };
 
 export const getPostsByAuthor = async (AuthorId: string) => {
-  try {
-    const { data } = await axiosInstance.get(`/posts/author/${AuthorId}`);
-    return data;
-  } catch (error) {
-    if (error instanceof Error) {
-      console.log(error.message);
-    } else {
-      console.log("Unknwon Error", error);
-    }
-  }
+	try {
+		const { data } = await axiosInstance.get(`/posts/author/${AuthorId}`);
+		return data;
+	} catch (error) {
+		if (error instanceof Error) {
+			console.log(error.message);
+		} else {
+			console.log("Unknwon Error", error);
+		}
+	}
 };
 
 export const createPost = async (post: Post) => {
-  try {
-    await axiosInstance.post("/posts/create", post);
-  } catch (error) {
-    if (error instanceof Error) {
-      console.log(error.message);
-    } else {
-      console.log("Unknwon Error", error);
-    }
-  }
+	try {
+		await axiosInstance.post("/posts/create", post);
+	} catch (error) {
+		if (error instanceof Error) {
+			console.log(error.message);
+		} else {
+			console.log("Unknwon Error", error);
+		}
+	}
 };
 
 export const updatePost = async (post: Post) => {
-  try {
-    await axiosInstance.put("/posts/update", post);
-  } catch (error) {
-    if (error instanceof Error) {
-      console.log(error.message);
-    } else {
-      console.log("Unknwon Error", error);
-    }
-  }
+	try {
+		await axiosInstance.put("/posts/update", post);
+	} catch (error) {
+		if (error instanceof Error) {
+			console.log(error.message);
+		} else {
+			console.log("Unknwon Error", error);
+		}
+	}
 };
 
 export const deletePost = async (postId: string) => {
-  try {
-    await axiosInstance.delete("/posts/delete", {
-      data: { id: postId },
-    });
-  } catch (error) {
-    if (error instanceof Error) {
-      console.log(error.message);
-    } else {
-      console.log("Unknwon Error", error);
-    }
-  }
+	try {
+		await axiosInstance.delete("/posts/delete", {
+			data: { id: postId }
+		});
+	} catch (error) {
+		if (error instanceof Error) {
+			console.log(error.message);
+		} else {
+			console.log("Unknwon Error", error);
+		}
+	}
 };
