+++ src/components/commons/Footer.tsx
@@ -1,7 +1,7 @@
 export default function Footer() {
-  return (
-    <>
-      <h1>Footer Component</h1>
-    </>
-  );
+	return (
+		<>
+			<h1>Footer Component</h1>
+		</>
+	);
 }
