+++ src/components/commons/ProtectedRoute.tsx
@@ -3,15 +3,15 @@ import { Outlet, useNavigate } from "react-router";
 import { useAuthStore } from "../../store/authStore";
 
 export default function ProtectedRoute() {
-  const [show, setShow] = useState(false);
-  const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
-  const navigate = useNavigate();
-  useEffect(() => {
-    if (!isLoggedIn) {
-      navigate("/login");
-      return;
-    }
-    setShow(true);
-  }, [isLoggedIn, navigate]);
-  return <>{show && <Outlet />}</>;
+	const [show, setShow] = useState(false);
+	const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
+	const navigate = useNavigate();
+	useEffect(() => {
+		if (!isLoggedIn) {
+			navigate("/login");
+			return;
+		}
+		setShow(true);
+	}, [isLoggedIn, navigate]);
+	return <>{show && <Outlet />}</>;
 }
