+++ src/components/commons/PublicOnlyRoute.tsx
@@ -3,15 +3,15 @@ import { Outlet, useNavigate } from "react-router";
 import { useAuthStore } from "../../store/authStore";
 
 export default function PublicOnlyRoute() {
-  const [show, setShow] = useState(false);
-  const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
-  const navigate = useNavigate();
-  useEffect(() => {
-    if (isLoggedIn) {
-      navigate("/");
-      return;
-    }
-    setShow(true);
-  }, [isLoggedIn, navigate]);
-  return <>{show && <Outlet />}</>;
+	const [show, setShow] = useState(false);
+	const isLoggedIn = useAuthStore((state) => state.isLoggedIn);
+	const navigate = useNavigate();
+	useEffect(() => {
+		if (isLoggedIn) {
+			navigate("/");
+			return;
+		}
+		setShow(true);
+	}, [isLoggedIn, navigate]);
+	return <>{show && <Outlet />}</>;
 }
