+++ src/main.tsx
@@ -4,9 +4,9 @@ import { BrowserRouter } from "react-router";
 import App from "./App.tsx";
 import "./css/index.css";
 createRoot(document.getElementById("root")!).render(
-  <BrowserRouter>
-    <StrictMode>
-      <App />
-    </StrictMode>
-  </BrowserRouter>
+	<BrowserRouter>
+		<StrictMode>
+			<App />
+		</StrictMode>
+	</BrowserRouter>
 );
