+++ src/pages/NotFound.tsx
@@ -1,7 +1,7 @@
 export default function NotFound() {
-  return (
-    <>
-      <h1>NotFound Component</h1>
-    </>
-  );
+	return (
+		<>
+			<h1>NotFound Component</h1>
+		</>
+	);
 }
