+++ src/pages/PostCreate.tsx
@@ -1,7 +1,7 @@
 export default function PostCreate() {
-  return (
-    <>
-      <h1>PostCreate Component</h1>
-    </>
-  );
+	return (
+		<>
+			<h1>PostCreate Component</h1>
+		</>
+	);
 }
